LinGui: Improve widget layout
Add padding to GtkEntry.  On some themes, it looked very crowded.

Let height of x264 extra settings text box float with height of widgets that
it shares a grid with.  It's fixed height was causing weird layout.

Reduce border around GtkComboBox.  There are lots of combo boxes, so
this results in a significant reduction in the overall size of the window.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6101 b64f7644-9d1e-0410-96f1-a4d463321fa5