QSV: always use our own DTS.
Previously, the DTS provided by MSDK was used
if it was available and the frame rate was constant.

Our own DTS is indentical to MSDK's (when the
frame rate is constant), but MSDK will sometimes
provide 1 or 2 bogus DTS over the duration of a
file, causing the encode to abort (lavf is picky).


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6158 b64f7644-9d1e-0410-96f1-a4d463321fa5