decpgssub: partial fix for "multi-object" PGS subs
We were packaging multiple subtitle packets with the same pts and passing
down to rendersub which would then throw away all but the last sub with 
a given pts. So put all AVSubtitleRects into the same hb_buffer_t.

A patch to libav is also required to fix the multi-object problem.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6211 b64f7644-9d1e-0410-96f1-a4d463321fa5