libhb: fix mp4 post tagging problems
A couple common mp4 taggers (mp3Tag and iTunes) fail to tag files
that have the Nero chapter atom (chpl).  This atom was added when
we switched to libavformat for muxing.  I haven't found any players
yet that rely on this for chapters.  If we ever find any, we can
consider adding an option somewhere to enable chpl.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6269 b64f7644-9d1e-0410-96f1-a4d463321fa5