CLI: fix dimensions problem when both width and height are specified
The GUI will precompute both width and height and specify both on the
command line.  But then the CLI recomputes width based on height which
can result in the width being changed.

So, in non-anamorphic mode, disable keep_aspect so that width and heigh
are both kept.  And in other anamorphic modes, use custom ana when computing
dimensions so that both width and height are kept (i.e. recompute par).


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6369 b64f7644-9d1e-0410-96f1-a4d463321fa5