libhb: improve preview generateion when there are no IDRs
Puts the decoder into a mode that returns *all* frames.  I.e.
it won't just scan to the end of the file looking for an IDR.  Then
we decode multiple frames (up to 100) searching for an I frame. This way,
the reconstructed frame we use for a preview is more likely to be complete.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6430 b64f7644-9d1e-0410-96f1-a4d463321fa5